Restructure application to contain bot and app separations
