+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "babel -w src/ -d build/ -s",
-    "build-css": "node-sass -w src/scss/ -d -o build/public/css",
+    "build-css": "node-sass -w src/app/scss/ -d -o build/public/css",
     "start": "node ./build/server.js",
     "start-dev": "node ./build/dev_slackbot.js"
   },
@@ -27,7 +27,8 @@
     "ejs": "^2.4.2",
     "express": "^4.13.4",
     "moment": "^2.13.0",
-    "moment-timezone": "^0.5.4"
+    "moment-timezone": "^0.5.4",
+    "pg": "^5.0.0"
   },
   "devDependencies": {
     "babel-cli": "^6.9.0",
