+++ src/app/scss/main.scss
@@ -1,6 +1,6 @@
 $test: 12px;
 
 body {
-	font-color: grey;
+	font-color: black;
 	font-size: $test;
 }
