+++ src/bot/bot/index.js
