+++ src/bot/controllers/work_sessions/endWorkSession.js
