+++ src/bot/controllers/work_sessions/middleWorkSession.js
