+++ src/bot/controllers/work_sessions/startWorkSession.js
