+++ src/bot/lib/botResponses.js
