+++ src/bot/middleware/hearMiddleware.js
